category and sub category delete action protected if there are dependent product and sub categories

+++ controllers/subCategoriesController.js
@@ -1,4 +1,5 @@
 const SubCategories = require('../models/subCategoriesModel');
+const Product = require('../models/productModel');
 const CustomError = require('../utils/CustomError');
 const catchAsync = require('../utils/catchAsync');
 const APIFeatures = require('../utils/apiFeatures');
@@ -75,22 +76,36 @@ exports.updateSubCategories = catchAsync(async (req, res, next) => {
 });
 
 exports.deleteSubCategories = catchAsync(async (req, res, next) => {
+	if (!req.params.id) {
+		return next(new CustomError('Please provide a sub-category ID', 400));
+		// Yes, it's a valid ObjectId, proceed with `findById` call.
+	}
 	if (!req.params.id.match(/^[0-9a-fA-F]{24}$/)) {
 		return next(new CustomError('Invalid sub-category ID provided.', 400));
 		// Yes, it's a valid ObjectId, proceed with `findById` call.
 	}
-	const subCategory = await SubCategories.findById(req.params.id);
-	if (!subCategory) {
-		return next(new CustomError('No sub-category found with that id', 404));
-	}
-	const imagePath = path.join(__dirname, '..', 'public', subCategory.image);
-	fs.unlink(imagePath, async err => {
-		if (err) {
-			await SubCategories.findByIdAndDelete(req.params.id);
-			return next(new CustomError('Error deleting sub-category image', 500));
-		} else {
-			await SubCategories.findByIdAndDelete(req.params.id);
-			return res.status(200).json({ status: 'success', data: {} });
-		}
+	// Checking if the category has dependent products or not
+	const subCategoryHasProducts = await Product.find({
+		sub_category_id: req.params.id,
 	});
+
+	if (!subCategoryHasProducts) {
+		const subCategory = await SubCategories.findById(req.params.id);
+		if (!subCategory) {
+			return next(new CustomError('No sub-category found with that id', 404));
+		}
+		const imagePath = path.join(__dirname, '..', 'public', subCategory.image);
+		fs.unlink(imagePath, async err => {
+			if (err) {
+				await SubCategories.findByIdAndDelete(req.params.id);
+				return next(new CustomError('Error deleting sub-category image', 500));
+			} else {
+				await SubCategories.findByIdAndDelete(req.params.id);
+				return res.status(200).json({ status: 'success', data: {} });
+			}
+		});
+	}
+	return next(
+		new CustomError('There are products that depend on this sub-category', 400)
+	);
 });
